Support simple logging (env npm_pnpm_logger=simple)
