+++ lib/config.js
@@ -1,4 +1,5 @@
 module.exports = require('rc')('npm', {
   pnpm_concurrency: 16,
-  pnpm_store_path: 'node_modules/.store'
+  pnpm_store_path: 'node_modules/.store',
+  pnpm_logger: 'pretty'
 })
